chore: fix typo in qf funding form filename
